requirements.txt
requests
Pillow
python-dotenv
aiogram
pandas
moviepy
ffmpeg-python
